allow -e script evaluation
Unfortunately, I don't know of a way to get rid of the `=` in
`perl6 bin/007 -e='say("OH HAI")'` -- it seems to be required
by Perl 6's built-in argument handling.

Someone is welcome to hack around this and get it working
without the `=`, as long as we still retain the nice USAGE
message.