get rid of the Val::Quasi type
This necessitates the stuttering of an expression into a statement. Thanks to
the tentative conclusion in https://github.com/masak/007/issues/7, I expect
this stuttering to be temporary.

Also change the first test in t/syntax/quasi.t to expect something slightly
different from (the AST returned from) a quasi. I'm not sure we'll keep that
output at all, but for now it's easier just to change the test to expect what
it actually gets.