remove block literals
There are still blocks, and you can still use (paramterless)
blocks on the statement level. But you can no longer store a
block in a variable, or return a block. Blocks don't go in
expressions in the code, and they are not values at runtime.

If you want something like this, declare a subroutine, and
then refer to it by name.

Also removed some tests that didn't make sense without block
literals. Rewrote some other tests.

Closes #11.