get rid of .declare method hack
Instead of calling a `declare` method on all statements in a block
just entered, we now do declarations in the proper place, in the
parser.

This causes AST-based tests to fail, since these ASTs were never
parsed. We solve this by introducing a `check` sub that AST-test
utilities are expected to use.

This `check` sub is eventually expected to be used in connection
with synthetic Qtrees (which were created by painstakingly nesting
ordinary Qtree objects inside each other). Some things that would
normally be caught at parse-time could then be caught at check-
time. (For example: variable used but not declared; redeclaration
of parameter; expected an operator but got a boolean.)

Note that the `check` sub is not up for this yet, but there seems
to be a path from the current point to that point.

As all the `declare` methods went away, the `enter` method needed
to pick up a bit of the slack and manually refresh subs in its
scope.

Closes #5.