fulfill all the wishes on setting-wishlist
(Thus also deleting setting-wishlist.)

We've accrued a bit of a testing deficit by doing this.
Also a deficit in typechecking. Hoping to remedy those two
deficits as we get deeper into synthetic Qtrees.

Some of these are funny. For example, `int` and `str` do double
duty as coercers of actual in-program values, and as accessors/
destructors of Qtree nodes (representing the int/str literals).