rename s/grep/filter/
The name 'grep' is an affectation within the Perl language and
community, caused by such things as being almost the first among
the dynamic languages to have a function like this, and wanting
to harken to Unix names and ways. After that, it's just inertia
that makes it keep this name. The name originally stems from an
'ed' command, and is an acronym for "Globally search a Regular
Expression and Print". Not much of this is relevant for the
modern function of selectively retaining elements from a list.

007 is not bound by tradition, inertia, or nostalgia. Hence we
switch to the Python name 'filter'.