introduce a Q::Compunit
Conceptually, a Q::Compunit has to be the root of any valid program.

Q::Compunit is a kind of Q::Statement::Block (in that it's this
"invisible block" that surrounds the whole compilation unit). But
since it's not an *actual* block, it felt wrong for
Q::Statement::Block to consist of a Q::Literal::Block, as it used
to. Refactored it to consist of a Q::Statements instead. Since
we still want a block to actually be there, a little more manual
work now falls on the clients of Q::Statement::Block to create
and enter one.

Addresses #20 but doesn't close it yet, because we've yet to move
the static lexpads up one level, which is possible now that we have
the compunit outermost level.