turn truthy() into methods
The problem in #21 was largely because truthy needed to follow/
cover the Val hierarchy, but there was nothing in place to
enforce or mandate that. Solution? Put the thing as methods on
the Val types themselves.

The whole thing is reminiscent of "feature envy", it's just that
the envy didn't emanate from another class, just from a set of
multis in a module. The refactor was basically a move into the
hierarchy we were mirroring. In the future if someone forgets to
add a truthy method we'll inherit the default always-True one.

By the way, this also fixes the somewhat weird detail that truthy
was talking only about Val, but it was defined at the top of Q.pm.
Not wrong as such, but contributing to the feeling of feature
envy. (It was used later in Q.pm, in Q::Statement::If, but it was
also used in Runtime.pm, which is why it needed exporting.)

I did consider other names, such as .so or .Bool. Decided to
postpone the rename for now, since nothing felt exactly right.