allow indexing on non-identifiers
Apparently we hadn't tested array indexing enough. I discovered when
I was putting together an example for an issue that there was a
hidden assumption that we were indexing an identifier, not an
arbitrary expression. (The "arbitrary expression" case already worked
for sub calls.) Seems it's been that way since always.

Added a test so that we don't regress on this.