Started implementing unquotes
A quasi can be full of holes -- so-called "unquotes". An unquote
contains an expression that needs to evaluate to a Qtree, which
then gets spliced directly into the code in the quasi. It's
quite similar to qq string interpolation, but with code instead
of a string. The unquote is a way of "escaping" out of the literal
code and into the land of Qtrees.

As an example, these all mean the same thing:

    quasi { say("Mr Bond!") }
    quasi { say({{{Q::Literal::Str("Mr Bond!")}}}) }
    my s = Q::Literal::Str("Mr Bond!"); quasi { say({{{s}}}) }
    my t = quasi { "Mr Bond!" }; quasi { say({{{t}}} }

(The last one doesn't work yet, I think. It should, though.)

The quasis are replaced by actual Qtrees at *evaluation*, when the
runtime "hits" the quasi by evaluating an expression in which it
is a part. At that point, we run down the Qtree structure of the
quasi block, cloning everything we find (except trivial leaf nodes,
which are safe to re-use), and replacing every unquote we find with
its expression (hopefully a Qtree). That's what the new .interpolate
method is for.

Addresses #30. Still a lot to do.