simplify Q::Statement::My structure a bit
After looking at IntelliJ's Java PSI, I realized that theirs is
a preferable way to implement Q::Statement::My and
Q::Statement::Constant. It looks like an assignment, but storing
it as one is redundant.

(It always did feel like a code smell and a OAOO risk that we had
to mention the variable name twice in those. Risk now averted;
smell febrezed.)