grand rename: -s to -list for stmts/params/args
I saw IntelliJ's PSI do this, and I liked it better than the
Perl 6 convention of having a Statement container be called
Statements.

Now that I think about it, having a plural for a class name
feels a little bit like a design smell. It's a single thing, but
it's naming itself as several things without saying exactly how.

List is much less ambiguous in this way. And all of the three
renamed things really are lists, in the sense that order matters
(as opposed to a set). Syntactically they're even written as
lists, even though statements happen to be semicolon-separated.