extract Q::Expr::Block from Q::Block
The former is now the strange weird thing generated as a result of
evaluating a quasi. Thanks for the naming suggestion, #7!

I first tried making Q::Expr::Block a subrole of both Q::Block and
Q::Expr, but that ended in hilarity, tragedy, and horror. First I
got a number of diamond-type conflicts with methods from Q, and
after uselessly fixing these, I got this wonderful error:

    No concretization found for Q

So for better or worse, Q::Expr::Block is currently a Q::Block but
not a Q::Expr.