also wrap values of List type
I don't remember exactly the situation in which I got a List
value and it failed to wrap correctly... it was at some point
while writing one of the example scripts. Anyway, this seemed
the obvious fix and it got me unstuck at the time.

Maybe there is some common type (like Iterable) that we should
check for instead. But I see in the Rakudo source that Map is
Iterable, despite being an associative type... so right now it
feels risky to do that, rather than just list all the known
concrete types.