s:g/OpLevel/OpScope/
Took some time to find a good name for this abstraction.

"Op level" sounds like it's to do with precedence levels. But
it's actually the thing that *holds* all the precedence levels
currently in play. "Op scope" captures it much better, since
this thing comes into being at '{' and gets destroyed at '}'.