define operators as built-ins
In df78e5aa, we moved all the operators out into the parser. (And
with 077e6fda, this happened at parser construction.) This was
requested by #29, in order to promote OAOO.

We (royal "we") are now making the equal and opposite decision:
defining the operators as built-ins, and having the parser ask
for an op scope to be created from what's there.

Some consequences of this:

- Op scopes are no longer parser-only. So maybe they should be
  moved out of the _007::Parser namespace into _007? (In general,
  the runtime doesn't have any dependencies on the parser, so
  it would make sense.) It's hard to say what an op scope "is" --
  whether it's genuinely something parser-related or not. (Some
  part of me says it is, as its main purpose is to allow decisions
  to be made about precedence/associativity.)

- Haven't tried it as part of this commit, but we're now just one
  step away from moving all the operator *logic* into the built-ins
  as well. This will make the operator Q types very boring. :)