move much of the operator logic into built-ins
As part of this, the Q::Prefix::Custom (et al.) types are gone.
There's no use talking about "custom" types in a system where
basically everything is user-defined.

I thought that as part of this commit, the [$type] role parameter
could go away. But it turns out this is how we do lookup in the
Q -> sub direction, so it stays (and moves up to the top op types).

It seems some parts of the 007 internals liked thinking of $type
as something looking like "[==]" (with the surrounding brackets)
and other parts as "==" (sans them). I didn't delve all too
deeply into why. Instead I just made things work for now. In the
long run, it might be good to arrive at a better kind of
consistency there.

Current exceptions -- operators that were not moved and that
instead override their .eval method:

- infix:<=> -- which has has "macroish" behavior, as it does not
  evaluate its left side into an rvalue, instead using the *name*
  of the identifier on the left side. Currently it seems to me
  that a builtin macro and a new `assign(name, val)` built-in sub
  could do the job. But need to test this.

- The three magical postfixes postfix:<[>, postfix:<(> and
  postfix:<.> -- no real difficulty here; they're not "macroish"
  as far as I can see. But I know they're a bit special-cased
  in the parser, so I saved them for now.