apply the X:: exception to the other obj lookup
As requested in
https://github.com/masak/007/commit/d6f18a89a147d0d79be7bb1ee1ff0d1e9f3e89fc#commitcomment-14387777

Note that there'll likely be a *third* such place when we implement
Object.get() -- this tells me that maybe this logic should be
pulled out to a separate place -- _007::Runtime, maybe -- and then
called from all 2-3 places.

Come to think of it, we would do well to have an overall plan for
object lookup in general, just to make sure Val::Object objects as
well as Q objects behave consistently.