refactor: remove Q:: .new methods with positionals
This in order to better close the loop between the Perl 6 level and
the 007 level -- as we move toward being able to create Q nodes from
inside of 007 userland, the positionals-based .new methods were
simply in the way. 007 creates objects in a key/value fashion, not
using positional parameters.

As part of this, vendethiel++ and I realized that the only place that
really uses the positional constructors is the _007::Test module and
its &read subroutine. Positional constructors made a lot of sense
when the only client to the Q nodes was the Lispish AST format, which
is positional in nature. Then the rest of the language was built on
top of that foundation, using the positional constructors without
really benefiting from them. In fact, adding in the nameds everywhere,
I had the sense that the source got a little bit clearer in places.
No doubt there are further wins to be had by clever variable naming
etc.

Some consequences of this:

- All the .interpolate methods now have to call the named .new methods
- Ditto all the constructor calls in the parser
- Ditto the built-in Q constructor functions
- Ditto some parts of _007::Test
- In general, things that used Empty as a marker for "child node not
  present" now use Any instead -- this being the default if nothing is
  passed into the constructor. (This currently happens in two places
  in the language: assigning nothing in a my/constant declaration, and
  returning nothing in a return statement. Notince that in both of
  these cases, there's a slight distinction between assigning/returning
  nothing at all, and assigning/returning None.)