Val:: refactor -- make all the Q types 007-native
Reason: now that it's becoming increasingly realistic to create
Q values from within 007, all the properties of these Q values
need to also be native to 007. More exactly, all the Val:: types
can still be opaque and have Perl 6 internals, but all the Q::
types must be built from only 007 stuff.

All in all, things have gotten a fair bit longer and more repetitious.
Some of that is extra clarity, which I like. Some of it doesn't really
add anything and ends up feeling unnecessary. We should look into
maybe introducing convenience functions/methods to soak up some of that
repetitiveness.

At some point, we introduced `handles` as a convenience to be
able to loop over statementlists, argumentlists, and parameterlists
more easily whenever they occur. Switching this over to Val::Array
means I backed out of using `handles`.

    - has @.properties handles <elems Numeric Real list>;
    + has Val::Array $.properties = Val::Array.new;

Which means many of the necessary changes end up looking like this:

    - my $count = $c.parameterlist.elems;
    + my $count = $c.parameterlist.parameters.elements.elems;

We could re-introduce `handles` and have it work "all the way down" to
Val::Array and its @.elements array. But I chose not to, at least for
now. There's a point to making the code non-magical and having it be
something that could easily be turned into 007 once it's time to
implement runtime and parser in 007.

In making such a port, we'll probably discover that a lot of the
"extra" method calls introduced in this commit go away again. This is
because 007 would already view the world as Val:: objects, whereas
the Perl 6 code needs to constantly peel away those to get at the
Perl 6 values it prefers.

Occasionally, we observe cases where the code actually gets
easier. For example, among the built-ins, we can now use the
incoming parameters directly, instead of peeling away a Val:: layer:

    - 'Q::Identifier' => -> $name { Q::Identifier.new(:name($name.value)) },
    - 'Q::StatementList' => -> $array { Q::StatementList.new(:statements($array.elements)) },
    - 'Q::ParameterList' => -> $array { Q::ParameterList.new(:parameters($array.elements)) },
    - 'Q::ArgumentList' => -> $array { Q::ArgumentList.new(:arguments($array.elements)) },
    + 'Q::Identifier' => -> $name { Q::Identifier.new(:name($name)) },
    + 'Q::StatementList' => -> $array { Q::StatementList.new(:statements($array)) },
    + 'Q::ParameterList' => -> $array { Q::ParameterList.new(:parameters($array)) },
    + 'Q::ArgumentList' => -> $array { Q::ArgumentList.new(:arguments($array)) },

As part of this refactor, a lot more parameters have been annotated
with types. This turned out to be a simple way to intercept values of
the wrong type and catch them at earlier places. I left the type
annotations in there.

Another funny detail was that string blackslash un-escaping, which belongs
with creating a new Val::Str, was moved from Q::Literal::Str.eval to
the parser (and a corresponding place in Test.pm's read()), because
Q::Literal::Str now ends up containing a Val::Str, and so it needs to be
created earlier.

Finally, I now believe the few `try` directives we have when looking for
macros are dangerous and should go away. The ".value" call that was added
in this change was necessary to make get-var find the right value (because
the .name in a Q::Identifier is now suddenly a Val::Str):

    - && (try my $macro = $*runtime.get-var($/.ast.name)) ~~ Val::Macro {
    + && (try my $macro = $*runtime.get-var($/.ast.name.value)) ~~ Val::Macro {

The presence of the `try` in there masked the exception that was being
thrown by .get-var (a X::TypeCheck::Binding exception, because it was
getting a Val::Str, not a Str). This would have been a useful exception
to get, rather than just silence. I've fallen into the usual trap of
"the exception is always the kind of exception I expect". We should
replace the `try` with a dedicated .get-var variant that fails in a nicer
way, indicating that nothing was found.