allow creating Val:: objects with obj term syntax
Note that this has to be special-cased by translating the (007)
Val::* value being passed in as a parameter (in the case of
Val::Int, Val::Str, and Val::Array) into the corresponding Perl 6
value. The option of turning the insides of Val:: objects
entirely into 007 types isn't available to us like it was for Q
objects, due to circularity constraints.

Also note that this is an excessively silly way to create any of
the built-in Val::* types. We're mostly supporting it out of a
sense of unfailing consistency.