Switch from using Any to Val::None.new in "empty" Q slots
Constructs such as Q::Statement::My and Q::Statement::Return have
an "optional" second slot. (See #84.) Up until now, we've marked
the absence of a Q value there by using Perl 6's Any type object.

This doesn't work if we want Q objects to be entirely internal in
the sense of commit 91bf49837a2e9b5b48bf264d1480f81e90629ca3. It's
not possible yet, but after #84 is fixed, the .expr property on
those nodes needs to contain a 007 value, and Any ain't it.

An argument could be made that "the absence of something here"
should be a Q type of its own. (Q::None or Q::Void, say.) We may
still take that path if other reasons for it show up. At this
point, we might as well re-use Val::None, which already has the
semantics of "you expected something here but there's nothing",
same as unset variables, default or absent `return` values, failed
lookups (though 007 throws an exception) and optional parameters
with no value passed to them (which 007 doesn't support yet).