eliminate `try` from all variable lookups
A `try` catches all kinds of unexpected failures, not just the one
having to do with not finding a variable. More to the point, in
all the cases where we were silencing the X::Undeclared exception,
we were doing so only because we wanted the *other* behavior in
the .get-var method: lookup which always succeeds.

If Perl 6 had monads, I might return the result as a Maybe[Val].
Instead, let's split up the two methods into four:

    $runtime.get-var
    $runtime.maybe-get-var

    $runtime!find
    $runtime!maybe-find

Where the original methods still throw an X::Undeclared, the
maybe- variants just return Nil.

    "No! Try not. Do. Or do not. There is no try." -- Yoda