place the declarator mutability info on Q types
Q::Statement::My and Q::Parameter produce symbols in the lexpad
that you can assign to. Used to be when we saw a `=` operator,
we special-cased 'my' and 'parameter' locally, in the code.
That feels non-extensible, and so we introduce the new abstract
Q::Declaration type instead.

A Q::Declaration is by default *non*-assignable, which feels
like a safe default. The above two declarations are exceptions,
and so they override the is-assignable method.

Currently we have five node types which are declarations:

- Q::Statement::My
- Q::Statement::Constant
- Q::Statement::Sub
- Q::Statement::Macro
- Q::Parameter