make a lot of Val and Q things classes
Instead of roles.

The rule is now: something that we plan to instantiate is a class.
Abstract types (that we only inherit from) are rules.

The demands on classes are a bit higher. You can't stub a class
and then declare it in a later module. Therefore this commit
also removes the stubbing, and adds in the Q::ParameterList and
Q::StatementList in all the .new calls.

For bootstrapping/stubbing reasons, the commit also moves
Val::Sub::Builtin into Runtime::Builtins.