Base the operators on identifiers, not strings
All operators now have an $.ident. Sometimes they override this
with an .ident method.

Custom operators are now simply normal operators with an $.ident
injected at construction. Instead of being passed around as type
objects until they're reduced in the expression parser, they're
now passed around as instances with only $.ident set. This
attribute is then manually cloned over to the "actual" instance.

I'm not very tied to this way of solving things, but it works.
It also has the merits of being fairly directly portable to 007.

Because Q::Postfix::Property already had an $.ident, I renamed
this to $.property instead.

Closes #82.