remove redundant XXX comment
The error message this comment wanted to introduce already exists,
but on the Actions side of things.

It feels slightly better to put a side effect such as a missing
initializer error message in the actions, actually. It simplifies
the parser, and it moves out logic into the actions. Generally,
we should only put logic and side effects in the grammar if they
absolutely need to happen in the middle of a rule parse.