refactor postfixes to be passed around sanely
We used to create a weird kind of tuple in some cases, like a
toolkit with instructions for building an actual postfix object.

This is no longer tenable as we now need to expose the partial
result returned by the postfix parsing rule. In other words,
this refactor is necessary -- besides being a generally good
idea.