do object property lookup in an automated way
Instead of a spotty, haphazard, inconsistent, and repetitive way.

This removes reams and reams of code the coverage tool reported as
being uncovered. (Though there are surely still some tests that
could be writte against every combination of Q type and method,
just to have some model coverage of that.)

The change also goes some way to fixing #81. The column "Built-in"
is now by necessity aligned with the column "Q attribute". Which
only leaves three effective columns to bring maximal consistency to.