lift out equality comparison mechanism from infix:<==>
Since it's now shared between infix:<==> and infix:<!=>, it feels
sensible to give it a wider shared scope, rather than letting
infix:<!=> call into infix:<==>.

This has a couple of advantages. First, we don't have to declare
a variable for infix:<==> so that we can call into it. Second, we
can compute non-equality entirely in Perl 6 code, without
coupling to the infrastructure we just set up in infix:<==>.

infix:<==> now ends up being a lot shorter, which is another small
advantage.