expose the 'ident' property of operator Q types
I feel a little conflicted about this one, but let's try it this
way.

The reason I feel conflicted is that, for the derived operator
types (currently, the ones corresponding to built-in operators),
the 'ident' property is "fixed" (known beforehand per Q type)
and therefore unnecessary. But it still needs to be specified,
and can be specified wrongly. "Just never do it wrong" doesn't
seem like a satisfactory answer in a system where the user is
able to create those Q objects synthetically.