consolidate 'interpolate' logic
This is one of those rare cases where logic *shouldn't* be spread out
as methods on respective types, but should instead be one single
recursive subroutine on the responsible type.

Various clues: the code in all the types was boring and repetitive,
and kept getting out of sync. It could be easily summarized with a
few lines of generic code. Unlike 'eval', which really does differ
between Q types, 'interpolate' mostly did the same thing all the
time: clone the object and its properties, giving something newish
enough that we can play with. Only with Q::Unquote was the
interpolation process even remotely interesting.