remove :assoc<right> on postfixes
These declarations would only matter if there were prefixes defined
on the same precedence level. There aren't any.

Of course, someone could add a user-defined prefix on the same
precedence level as one of the existing postfixes. But it's not
clear to me in that case that :assoc<right> would be a saner
behavior than :assoc<left>. (I think it'd be confusing either
way.)