implement prefix:<^>
This one came out of the left field, I know. But I was looking at
hanoi.007 in the (still very much a work-in-progress) examples/
branch, and I saw that we implemented our own `upto` function.

Then I looked at the ideal version of that script that I had
produced a few weeks ago, and it had a prefix:<^>. That settled it.