remove dependence on GNU Find
Ideally, we'll want to not depend on any platform-dependent commands.
But for now, we'll handle these on a case-by-case basis as they are
reported as causing problems for people.

In this case, I lifted in File::Find
(<https://github.com/tadzik/File-Find>, MIT-licensed), and then
modified it to only handle what we needed in this case.

Fixes #121.