make "update" and "extend" work on the old object
Decided to still make them return the "new" object, so you can chain
things. It's just that the "new" object is now the old object (i.e.
same object reference) that's been modified.

I got the eerie feeling when I was modifying these two methods, that
we don't do any checking of the inputs, and we probably should. I.e.
the `update` method is meant to assert that only old properties are
accepted, and the `extend` method is meant to accept only new ones.
That's why we have two methods in the first place.