add regression test for if statement parsing
We discovered this breakage while rebasing the branch where we
were doing the format macro. Adding the test -- which passes --
just so we don't regress on it again.

The regression was caused by turing `token statement:if` into
`rule statement:if`, which made the rule look nicer, but changed
the whitespace gobbling behavior at the end of the rule.

moritz++ points out that <.ws> should only consume insignificant
whitespace, not whitespace which might be instrumental in
determining the termination of a statement. jnthn++ points out
that Rakudo has an <?ENDSTMT> mechanism for this. Both of these
are good cues for what to do next so that we can actually nicely
turn `token statement:if` (and others) into a rule.