fix unbalanced block finishing for pblocks
Discovered when writing real-world examples and getting spurious
errors.

Short story: not only do `newpad` and `finishpad` have to match
up always, but `newpad` and `finish-block` (in Actions) also have
to match up. They are further apart, so it's harder to see when
they don't.

An easy rule of thumb though is that whoever creates the Q::Block
is the one that should also call `finish-block`. This is what
the <pblock> rule didn't do in this case.

A quick scan reveals that we do it in all other cases -- except
in `property:method`. Will fix that in a separate commit.