add new test file t/integration/finishblock.t
We already have this test file in master, but this one does the
same test using QAST, not sloppy regex matching.

Not quite done yet, though. The test as it stands is more or less
unreadable, and exposes a lot of QAST internals that we'd be
better off hiding away in utility routines. Let's try that in
this branch.