implement unexpanded mode
This simple implementation delivers the right tree, and makes sure
that the macros in question still *run*. (So you will see side
effects that happen from them, etc.)

This is fine for now; in the fullness of time I can imagine that
this will not be enough. Specifically, we might be interested in
keeping the expanded tree around somewhere for Sufficiently
Advanced Macros to look at. With this simple solution, we just
render and return the unexpanded tree. It doesn't hurt us right
now because macros can't introspect the program tree being built
yet.

Closes #64.