consolidate all stringification into one helper sub
The helper sub needs to be an `our` sub in a class, so that it
can be reached from _007::Q too. I would have preferred to
declare it in lib/_007/Q.pm, but apparently then I can't
predeclare it in lib/_007/Val.pm. So this appears to be a
reasonable compromise.

Closes #137.