back-paddle on the `None {}` syntax
As part of b2f7dbc3117ccad2fb363e32c1020585c3a80b8d, it was decided
that `None {}` should be used to "create" new None instances. The
reasons at the time were felt to be... consistency, mostly. You can
do it with all the other Val:: types, so why not with Val::None?

However, the following code shows why this is a bad idea:

    if qnode != None {
        return;     # parse error at `;` -- not part of
                    # the object literal syntax
    }

Yes, this is the classic "block gobbling" of Perl 6.

I thought this would be a minor thing, and that the consistency
of being able to instantiate `None` like this would win over the
inconvenience of block gobbling. But after trying things out, it
turns out it's the other way around.

Anyway, no big loss. We can "special-case" `None` to not be
instantiable, in a way that does not hurt at all. Because `None`
is essentially a singleton type (where `None` in fact represents
both the instance and the type), since it has no data/state, and
all instances are indistinguishable under the `==` relation.