add some abstract base classes
Was playing around, and tried to run this 007 program:

    my q = Q::Literal::Int { value: 42 };
    say(q ~~ Q::Literal)

But it came back saying that `Q::Literal` wasn't a thing, which
felt wrong. So I added them. Now the above program prints "1\n".

We might have to make some extra effort to make those abstract types
uninstantiable. But that can be a separate commit.