introduce the `new` keyword
So now 007 has more of a "normal" syntax in that regard. :-/

When supplying the type of object to create, the `new` keyword
now has to be used:

    new Q::Identifier { name: "foo" }       # fine
    Object { yay: "nay" }                   # fine

However, when just creating an `Object` without specifying its type,
the `new` keyword is allowed but optional:

    { yay: "nay" }                          # fine
    new { yay: "nay" }                      # fine

The old syntax, supplying the type of object but without the `new`
keyword, is now an error:

    Q::Identifier { name: "foo" }           # ERROR
    Object { yay: "nay" }                   # ERROR

(This commit does not add a special parse error to catch this old
form of object construction. That's still a good idea.)

Efforts were made to update all of the test suite, the examples,
and the self-hosting directory to use the new syntax.

Closes #147.