pass Qtree in intact, instead of serializing
Used to be the outer runtime sent the Qtree into the inner
runtime by injecting the source text for the appropriate Qtree
into the source code of the inner runtime.

On reflection, this is overly indirect. We have access to the
static lexpads of the inner runtime from the code that passes
it into the outer runtime; at that point, we just tweak the
`ast` variable to contain the generated Qtree.

One advantage of this is that Runtime.pm and runtime.007 are
now playing with exactly the same Qtree objects.