collapse a test into the actual implementation
After #141, the test actually became meaningless without us quite
realizing it.

What does the test check for? It checks that there's a consistency
between the Val.pm and Q.pm files on the one hand and the
Runtime/Builtins.pm file on the other. But the #141 upgrade
actually provides a way for the code itself to ask this question,
and automatically keep things consistent. So it shouldn't be a
test anymore, it should be the code.

And now, it is.