do a maybe-get-var check for undeclared variables
When we see an undeclared variable, we postpone a check to the
end of the compunit compilation that re-checks whether that
variable is now a sub (in which case it's OK). But we should
do a `maybe-get-var` to look up the variable, since the
assumption is that it's still undeclared, and we prefer to
fail in our "own" code, not somewhere inside `get-var`.