Get rid of Val::Sub::Builtin
This type was introduced out of necessity -- we've had it all the
way back to 5de4d04b64a9d5f1f5ec91f72c619d8ba9eb0bcf -- because
the need was felt to "extend" `Val::Sub` with builtin functionality.

This commit does everything with one single type -- `Val::Sub` --
but introduces a new field `&.hook` which built-in subs set.

As an added bonus, various special cases that needed to deal with
`Val::Sub::Builtin` disappear from the code, notably:

- When getting the type name, we no longer need to manually remove
  the `::Builtin` part to make `Val::Sub::Builtin` pretend to be
  a `Val::Sub`.

- We no longer need to skip `Val::Sub::Builtin` from declaring
  itself among the 007 built-in types.

As an added bonus, nowadays when we declare built-in subs and
methods in various places, we can do it with shorter syntax,
because we have more control over how it gets created. Most
importantly, we can often lift the name of the sub from various
outside circumstances.

(Though I do note there are still things we can do more
cleverly in the `_007::Runtime.property` method. There's a bit
too much repetition and runtime generation in there. Also, it
seems it's perhaps doing things that `_007::Runtime::Builtins`
should be responsible for.)

Also a nice side-effect: we no longer have to make
`$.parameterlist` read-writable, because we can now create it
with the correct parameterlist from the outset.

I tried to look for references where people assert it's a bad
idea to use inheritance in this type of case, and that using
attributes (possibly hooks) is preferable, but I didn't find
any. Would still be interested in finding such references.