Convincingly lie to the user about the op names
Internally, the name is definitely `infix:+`. But whenever 007 is
asked to print this name, it will print `infix:<+>`. Some more
rules apply to make the choice of <> vs «» and backslashes come
out right.

The overall idea, of course, is that the name should round-trip:
the same name that 007 prints could also have been used to define
the operator in code. (We don't guarantee that the operator *was*
defined in that particular way, only that it could have been.)

Closes #117.