allow `{{{op @ Q::ArgumentList}}}`
This also starts making use of an optional parameter to the
`unquote` rule, passing along the expected type in the context
so that we can fail earlier, inside of the rule instead of
outside it.

This will go some way to getting out of the quagmire described in
https://github.com/masak/007/issues/30#issuecomment-214724263 --
it was enough for Q::ArgumentList, for example, which clashes with
term position -- but I'll be quick to note that it doesn't
address the problem at its root. Notably, when we do hit the type,
at the end of the unquote, we've already had who-knows-how-many
side effects along the way from the stuff inside that we parsed.

In other words, something like `unquote(qtype @ expr)` would
still be beneficial, maybe even necessary, in order to finish up
issue #30.