Implement Q::CompUnit unquotes
A little bit unsettling that the Q::CompUnit node can't be generated
in a "natural" setting, since there can only even be *one* compunit
in a piece of program text, and that node has already been "spent"
at the time we are able to unquote. Instead we just check that we
can pass in a Q::CompUnit produced by another quasi.