Implement the 'class' statement
This is only a first stab; you can put statements inside of the
class body, but they won't run; nor will you be able to create
instances of the class yet.

This is the first feature that is developed behind a feature flag.
We'll see how it goes. See
http://martinfowler.com/articles/feature-toggles.html for more
information.